Add the ability to move images in the first section while hover
each image has 2 states , the origin coordinates state and the offset x and y state , where the first stores the coordinates of the origin of the image with respect to the viewport , and the second changes as mouse moves over the image according to the difference between the new mouse position with respect to the viewport and the origin of the image.
To be done : extract the image into a separate component to be reusable.