Create slider CardsSlider in RecommendationsErrorPage
the slider works by having a window with width equal to the details container's width , and a flex container of products that moves.
the slide happens by turning on the transition , sliding to the next slide using transform, turning off the transition , updating the products list to have the first product in the end and then removing the transform so that it's not displayed that there was a transform (translate) from product 2 to 3 and then 3 to 2. now one is at the end , 2 is displayed and 3 is after 2. repeat endlessly